♻️ [Refactor]: study room layout 컴포넌트 이동
