+++ src/features/study-room/component/StudyRoomIntendedCard.tsx
